feat(nice-grpc): include support for implicit secure protocol
If a client is set up using secure credentials with only the hostname
and no protocol, the current logic defaults to the insecure 80 port.

This refactors the createChannel logic, and then adds some logic to take
into account the possibility that the credentials passed in are already
set up to be secure.

Since ChannelCredentials is an abstract class - we are always going to
be passed a credential object that has _isSecure() set to either true or false
(or nothing at all, which is then default insecure).

This also means we shouldn't need to re-call createSSL or createInsecure
unless the credential object is missing completely.