fix: preserve extensions in server reflection
This commit updates the server reflection implementation to be lossless by default.

The current server reflection implementation was lossy: it was using a decode/encode step to generate encode file descriptor. This operation does not round-trip by default: all unknown fields are lost. In particular, all option extensions are dropped. Option extensions are often used as a way to attach metadata to Protobuf descriptors so this lossy behavior prevents compat with more advanced reflection use cases. [See this comment for details about the issue](https://github.com/deeplay-io/nice-grpc/pull/811#issuecomment-3204494409).

The `ServerReflection` function now uses slices of the original input to preserve all data.

This commit also marks the `serviceNames` option as readonly to match how it is used.